crm_api/requirements.txt
-i https://pypi.org/simple
atomicwrites==1.3.0
attrs==19.1.0
django-cors-headers==2.5.2
django==2.2
djangorestframework==3.9.2
gunicorn==19.9.0
more-itertools==7.0.0 ; python_version > '2.7'
pluggy==0.9.0
py==1.8.0
pytest-django==3.4.8
pytest==4.4.0
python-dotenv==0.10.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
-i https://pypi.org/simple
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
django-cors-headers==2.5.2
django==2.2
djangorestframework==3.9.2
drf-yasg[validation]==1.15.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
jinja2==2.10.1
jsonschema==3.0.1
markupsafe==1.1.1
more-itertools==7.0.0 ; python_version > '2.7'
pluggy==0.9.0
psycopg2==2.8.2
py==1.8.0
pyrsistent==0.14.11
pytest-cov==2.6.1
pytest-django==3.4.8
pytest==4.4.0
python-dotenv==0.10.1
pytz==2019.1
pyyaml==5.1
requests==2.21.0
ruamel.yaml==0.15.91
six==1.12.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
-i https://pypi.org/simple
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
django-cors-headers==2.5.2
django==2.2
djangorestframework==3.9.2
drf-yasg[validation]==1.15.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
jinja2==2.10.1
jsonschema==3.0.1
markupsafe==1.1.1
more-itertools==7.0.0 ; python_version > '2.7'
pluggy==0.9.0
psycopg2==2.8.2
py==1.8.0
pyrsistent==0.14.11
pytest-cov==2.6.1
pytest-django==3.4.8
pytest==4.4.0
python-dotenv==0.10.1
pytz==2019.1
pyyaml==5.1
requests==2.21.0
ruamel.yaml==0.15.91
six==1.12.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
